UT-RAG/requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
llama-index-llms-google-genai
llama-index-embeddings-google-genai
llama-index-vector-stores-pinecone
llama-index-readers-web
pinecone
slowapi
google-generativeai
beautifulsoup4
requests